requirements.txt
langchain-google-genai
python-dotenv
pydantic
langchain-core
langchain-text-splitters
langchain-community
langchain

# Local UI/API
fastapi
uvicorn[standard]
python-multipart

# Fast Ingestion
pymupdf

# Lightweight Vector Search & Embeddings
faiss-cpu
langchain-huggingface
sentence-transformers

# Hybrid Search
rank_bm25
